zshrc: avoid loop in SSH_AUTH_SOCK

+++ .config/zsh/.zshrc
@@ -40,7 +40,7 @@ fi
 export PATH="$_EXTRA_PATH:$PATH"
 
 # Update symlink for all tmux tabs
-if [ -S "$SSH_AUTH_SOCK" ]; then
+if [ -S "$SSH_AUTH_SOCK" ] && [ "$SSH_AUTH_SOCK" != "$HOME/.ssh/ssh_auth_sock" ]; then
     ln -sf $SSH_AUTH_SOCK ~/.ssh/ssh_auth_sock
 fi
 export SSH_AUTH_SOCK="$HOME/.ssh/ssh_auth_sock"
